requirements.txt
# Web Framework
Flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
gunicorn==21.2.0

# Deep Learning
tensorflow==2.13.0
keras==2.13.1

# Image Processing
opencv-python-headless==4.8.0.76
Pillow==10.0.0
scikit-image==0.21.0
scipy==1.11.2

# Utilities
numpy==1.24.3
